ui changes and restructuring
