+++ src/modules/arcade/tetris/tetris.js
@@ -1,11 +1,11 @@
 import { LightningElement, track } from 'lwc';
-import Engine from '../engine/engine';
-import Canvas from '../../view/model/canvas';
+import Engine from './engine/engine';
+import Canvas from '../../view/divCanvas/model/canvas';
 
-export default class App extends LightningElement {
+export default class Tetris extends LightningElement {
     // Note: the canvas (array) needs to be declared here, to fulfill all tracking requirements.
     @track canvas;
-    @track next;
+    @track nextView;
     engine;
     
     actions = {
@@ -22,8 +22,8 @@ export default class App extends LightningElement {
         this.engine && this.engine.stop();
         
         this.canvas = new Canvas(10, 20);
-        this.next = new Canvas(4, 4);
-        this.engine = new Engine(this.canvas, this.next);
+        this.nextView = new Canvas(4, 4);
+        this.engine = new Engine(this.canvas, this.nextView);
     }
     
     execute = (evt) => {
@@ -39,5 +39,13 @@ export default class App extends LightningElement {
     
     disconnectedCallback() {
         document.removeEventListener('keydown', this.execute);
+    }
+    
+    get level() {
+        return this.engine.level;
+    }
+    
+    get score() {
+        return this.engine.score;
     }
 }
