+++ src/modules/view/divCanvas/divCanvas.html
@@ -1,0 +1,13 @@
+<template>
+    <div class="canvas" style={canvasStyle}>
+        <div for:each={canvas} for:item="row" key={row.key} class="row" style={rowHeight}>
+            <div for:each={row} for:item="pixel" key={pixel.key} class="border" style={border}>
+                <div class="pixel"
+                     style={pixel.style}
+                     data-x={pixel.x} data-y={pixel.y}
+                     onclick={dispatchClick}>
+                </div>
+            </div>
+        </div>
+    </div>
+</template>
