+++ src/modules/view/divCanvas/model/shape.js
