--- src/modules/view/pixels/pixels.html
@@ -1,11 +1,0 @@
-<template>
-    <div class="canvas" style={canvasStyle}>
-        <div for:each={canvas} for:item="row" key={row.key} class="row" style={rowHeight}>
-            <div for:each={row} for:item="pixel" key={pixel.key}
-                class="pixel" style={pixel.style}
-                data-x={pixel.x} data-y={pixel.y}
-                onclick={rethrowClick}>
-            </div>
-        </div>
-    </div>
-</template>
